pyproject.toml
[project]
name = "taed2-2425q1-demos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "codecarbon>=2.8.3",
    "debugpy>=1.8.12",
    "deepchecks>=0.19.1",
    "dvc>=3.59.1",
    "fastapi>=0.115.8",
    "great-expectations>=1.3.6",
    "mlflow>=2.20.2",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "tensorflow-hub>=0.16.1",
    "tensorflow>=2.18.0",
    "torchvision>=0.21.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "pillow>=11.1.0",
    "pandas>=2.1.1",
]

[dependency-groups]
dev = [
    "bandit>=1.8.3",
    "deptry>=0.23.0",
    "httpx>=0.27.2",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.6",
]

[tool.coverage.run]
omit = ["src/prepare.py", "src/evaluate.py", "src/train.py", "src/train_api_demo_models.py"]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
addopts = "--junitxml=out/tests-report.xml --cov=src --cov-report=html:out/coverage"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    #Pylint
    "PL",
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true
